+++ AutoTrader/Trader/Repository/Lykke/PayloadDTOs/PayloadHistoryRate.cs
@@ -5,10 +5,10 @@ namespace AutoTrader.Repository
     public class PayloadHistoryRate
     {
         public string Id { get; set; }
-        public float Bid { get; set; }
-        public float Ask { get; set; }
-        public float TradingVolume { get; set; }
-        public float TradingOppositeVolume { get; set; }
+        public Decimal Bid { get; set; }
+        public Decimal Ask { get; set; }
+        public Decimal TradingVolume { get; set; }
+        public Decimal TradingOppositeVolume { get; set; }
     }
 
 }
